refactor some code, fix the way PRO_CLICKED is dispatched
